fixed issue where you couldn't tune frequency above 4.294 GHz

+++ cpp/record_iq_blade.cpp
@@ -54,7 +54,7 @@ int main(const int argc, const char *argv[])
 	bool saturated = false;
 	std::uint32_t overrunCounter = 0;
 
-	const std::uint32_t frequencyHz = atof(argv[1])*1e6;
+	const std::uint64_t frequencyHz = atof(argv[1])*1e6;
 	const std::uint32_t requestedBandwidthHz = atof(argv[2])*1e6;
 	std::uint32_t receivedBandwidthHz = 0;
 	const std::uint32_t requestedSampleRate = atof(argv[3])*1e6;
